Fixed warnings and tabs.

+++ ass1/ass1/Exercises.h
@@ -8,22 +8,22 @@
 std::pair<float, float> Statistics(const std::list<float>& values)
 {
     float avg = 0, stddev = 0, sum = 0;
-    int size = values.size();
-    
+    size_t size = values.size();
+
     if (size == 0)
         return std::pair<float, float>(avg, stddev);
-    
+
     for (float v : values)
         sum += v;
-    
+
     avg = sum / size;
-    
+
     for (float v: values)
         stddev += powf(v - avg, 2);
-    
+
     stddev /= size;
     stddev = sqrtf(stddev);
-    
+
     return std::pair<float, float>(avg, stddev);
 }
 //////////////////////////////////////////////////////////////////
